+++ SampleProject~/Assets/Samples/01 - Context Screen/Scripts/ContextScreenView.cs
@@ -24,7 +24,7 @@ namespace Unity.Advertisement.IosSupport.Components
 #if UNITY_IOS
             Debug.Log("Unity iOS Support: Requesting iOS App Tracking Transparency native dialog.");
 
-            ATTrackingStatusBinding.RequestAuthorizationTracking();
+            ATTrackingStatusBinding.RequestAuthorizationTracking(AuthorizationTrackingReceived);
 
             sentTrackingAuthorizationRequest?.Invoke();
 #else
@@ -32,5 +32,11 @@ namespace Unity.Advertisement.IosSupport.Components
                              "but the current platform is not iOS.");
 #endif
         }
+        
+        private void AuthorizationTrackingReceived(int status) {
+             Debug.LogFormat("Tracking status received: {0}", status);
+
+        }
+
     }
 }