+++ com.unity.ads.ios-support/Runtime/Plugins/iOS/Categories/UASupportTools.m
@@ -1,0 +1,10 @@
+#import "UASupportTools.h"
+
+
+_Nullable id uasupport_typecast(id obj, Class class) {
+    if ([obj isKindOfClass: class]) {
+        return obj;
+    } else {
+        return nil;
+    }
+}