+++ com.unity.ads.ios-support/Runtime/Plugins/iOS/TrackingAuthorizationManager.h
@@ -1,11 +1,13 @@
 #import <Foundation/Foundation.h>
 
+typedef void (*TrackingAuthorizationCompletion)(NSUInteger status);
+
 @interface TrackingAuthorizationManager : NSObject
 
 + (TrackingAuthorizationManager *)sharedInstance;
 
 - (BOOL)isAvailable;
-- (void)trackingAuthorizationRequest;
+- (void)trackingAuthorizationRequest:(TrackingAuthorizationCompletion)completion;
 - (NSUInteger)getTrackingAuthorizationStatus;
 
 @end