+++ com.unity.ads.ios-support/Runtime/Plugins/iOS/TrackingAuthorizationManager.m
@@ -15,9 +15,9 @@
 + (TrackingAuthorizationManager *)sharedInstance {
     static TrackingAuthorizationManager *instance = nil;
     static dispatch_once_t token;
-
+    
     dispatch_once(&token, ^{
-      instance = [[TrackingAuthorizationManager alloc] init];
+        instance = [[TrackingAuthorizationManager alloc] init];
     });
     return instance;
 }
@@ -36,14 +36,21 @@
     return self.trackingManagerAuthorizationClass != nil && [[NSBundle mainBundle] objectForInfoDictionaryKey:@"NSUserTrackingUsageDescription"] != nil;
 }
 
-- (void)trackingAuthorizationRequest {
-    if (!self.isAvailable)
+- (void)trackingAuthorizationRequest:(TrackingAuthorizationCompletion)completion {
+    if (!self.isAvailable) {
+        if (completion != nil) {
+            completion(0);
+        }
         return;
-
+    }
+    
     id handler = ^(NSUInteger result) {
-      NSLog(@"Result request tracking authorization : %lu", (unsigned long)result);
+        NSLog(@"Result request tracking authorization : %lu", (unsigned long)result);
+        if (completion != nil) {
+            completion(result);
+        }
     };
-
+    
     SEL requestSelector = NSSelectorFromString(@"requestTrackingAuthorizationWithCompletionHandler:");
     if ([self.trackingManagerAuthorizationClass respondsToSelector:requestSelector]) {
         [self.trackingManagerAuthorizationClass performSelector:requestSelector withObject:handler];
@@ -53,29 +60,29 @@
 - (NSUInteger)getTrackingAuthorizationStatus {
     if (!self.isAvailable)
         return 0;
-
+    
     NSUInteger value = [[self.trackingManagerAuthorizationClass valueForKey:@"trackingAuthorizationStatus"] unsignedIntegerValue];
-
+    
     return value;
 }
 
 + (BOOL)isFrameworkPresent {
     id attClass = objc_getClass("ATTrackingManager");
-
+    
     if (attClass)
         return TRUE;
-
+    
     return FALSE;
 }
 
 + (BOOL)isDeviceSimulator {
     struct utsname systemInfo;
     uname(&systemInfo);
-
+    
     NSString *model = [NSString stringWithCString:systemInfo.machine encoding:NSUTF8StringEncoding];
     if ([model isEqualToString:@"x86_64"] || [model isEqualToString:@"i386"])
         return TRUE;
-
+    
     return FALSE;
 }
 
@@ -85,7 +92,7 @@
         NSLog(@"AppTrackingTransparency Framework is not present, trying to load it.");
         if ([TrackingAuthorizationManager isDeviceSimulator]) {
             NSString *frameworkPath = [[NSProcessInfo processInfo] environment]
-                [@"DYLD_FALLBACK_FRAMEWORK_PATH"];
+            [@"DYLD_FALLBACK_FRAMEWORK_PATH"];
             if (frameworkPath) {
                 frameworkLocation = [NSString pathWithComponents:@[ frameworkPath, @"AppTrackingTransparency.framework", @"AppTrackingTransparency" ]];
             }
@@ -93,7 +100,7 @@
             frameworkLocation = [NSString stringWithFormat:@"/System/Library/Frameworks/AppTrackingTransparency.framework/AppTrackingTransparency"];
         }
         dlopen([frameworkLocation cStringUsingEncoding:NSUTF8StringEncoding], RTLD_LAZY);
-
+        
         if (![TrackingAuthorizationManager isFrameworkPresent]) {
             NSLog(@"AppTrackingTransparency still not present!");
             return FALSE;