+++ com.unity.ads.ios-support/Runtime/iOSTransparency/ATTrackingStatusBinding.cs
@@ -4,12 +4,15 @@ using System.Runtime.InteropServices;
 
 namespace Unity.Advertisement.IosSupport
 {
+    using AOT;
     public class ATTrackingStatusBinding
     {
 #if UNITY_IOS
-        [DllImport("__Internal")] private static extern void InterfaceTrackingAuthorizationRequest();
+        [DllImport("__Internal")] private static extern void InterfaceTrackingAuthorizationRequest(RequestAuthorizationTrackingCompleteHandler callback);
         [DllImport("__Internal")] private static extern int InterfaceGetTrackingAuthorizationStatus();
 #endif
+        public delegate void RequestAuthorizationTrackingCompleteHandler(int status);
+        private static RequestAuthorizationTrackingCompleteHandler _requestAuthorizationTrackingCompleteCallback = null;
 
         /// <summary>
         /// The enumerated states of an authorization tracking request.
@@ -27,11 +30,23 @@ namespace Unity.Advertisement.IosSupport
         /// </summary>
         public static void RequestAuthorizationTracking()
         {
+            RequestAuthorizationTracking(null);
+        }
+
+        /// <summary>
+        /// This method allows you to <a href="https://developer.apple.com/documentation/apptrackingtransparency/attrackingmanager/3547037-requesttrackingauthorization">request the user permission dialogue</a>.
+        /// </summary>
+        public static void RequestAuthorizationTracking(RequestAuthorizationTrackingCompleteHandler callback)
+        {
 #if UNITY_IOS
-            if (Application.platform == RuntimePlatform.IPhonePlayer)
+            if (Application.platform != RuntimePlatform.IPhonePlayer) return;
+            if (_requestAuthorizationTrackingCompleteCallback != null)
             {
-                InterfaceTrackingAuthorizationRequest();
+                throw new InvalidOperationException("App tracking transparency request is already triggered and awaiting completion");
             }
+            _requestAuthorizationTrackingCompleteCallback = callback;
+
+            InterfaceTrackingAuthorizationRequest(AppTransparencyTrackingRequestCompleted);
 #endif
         }
 
@@ -49,5 +64,13 @@ namespace Unity.Advertisement.IosSupport
 #endif
             return AuthorizationTrackingStatus.NOT_DETERMINED;
         }
+        
+        [MonoPInvokeCallback(typeof(RequestAuthorizationTrackingCompleteHandler))]
+        public static void AppTransparencyTrackingRequestCompleted(int status)
+        {
+            _requestAuthorizationTrackingCompleteCallback?.Invoke(status);
+            _requestAuthorizationTrackingCompleteCallback = null;
+        }
+        
     }
 }